shell: refactor builtin.py to util_rules
